rewrite commands
