+++ node/src/commands/update/index.js
@@ -1,0 +1,11 @@
+const schemaCommand = require('./schema');
+const objectCommand = require('./object');
+
+exports.command = 'update <schema|object>';
+exports.describe = 'Update CRM schema or object instance';
+
+exports.builder = (yargs) => {
+  yargs.command(schemaCommand);
+  yargs.command(objectCommand);
+  return yargs;
+};
